Makes AnnotationProcessor able to Binds interactorOutput with output Action Method
OutputAnnotationProcessor was generating code looking for The type of
field that was returned, That was not allowing to have two interactors
returning the same type of data in the same presenter. This
modification allows it. The only thing you have to do is concatenate
the name of the output in cameCase to the output annotated method as
you can see at this commit in DetailPresenter.java and
MainPresenter.java.